fix: keep binding manager below source size ceiling
Agend-Agent: archfix-codex-dev
Agend-Task: t-20260714145900416954-27634-127
Agend-Branch: fix/lifecycle-home-alias-release-truth
Agend-Issued-At: 2026-07-14T15:28:34.093486+00:00
